app/reporting: reindex inventory devices
on inventory trigger, pull and store/update the device.

we're using the same Device struct to both insert and update
(already knows how to serialize itself).

in case of updates, technically we should prepare a diff
between the existing ES device and incoming inventory API device, but
this proved to be too complicated for a quick POC. the whole document
is updated instead.

there is a followup task for 'really' implementing this endpoint,
leaving it until then.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>